Added the ability to send TLS alerts based on certain error conditions
Using this mechanism a python exception related to GNUTLS can be mapped to a
TLS alert and sent to the peer which will map it back to the original python
exception, making it possible to transfer error conditions and raise their
corresponding exception on the other side that becomes this way aware of the
errors that occured in the peer. Currently this is used to map certificate
related exceptions into TLS alerts and back to exceptions on the other side.